Move file to rest runners
