+++ Tests/TestRunners/SelfVerifyingTestFactory.cs
@@ -6,7 +6,6 @@ using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using CodeConverter.Tests.Compilation;
-using CodeConverter.Tests.CSharp;
 using ICSharpCode.CodeConverter;
 using ICSharpCode.CodeConverter.CSharp;
 using ICSharpCode.CodeConverter.Shared;
